refactor: rename textract_layout_chunker to layout_chunk_handler
