+++ tests/orchestration/test_pipeline.py
@@ -5,7 +5,7 @@ from unittest import mock
 import pytest
 
 from ingestion_pipeline.chunking.schemas import DocumentMetadata
-from ingestion_pipeline.chunking.strategies.layout.textract_layout_chunker import ChunkError
+from ingestion_pipeline.chunking.strategies.layout.layout_chunk_handler import ChunkError
 from ingestion_pipeline.embedding.embedding_generator import EmbeddingError
 from ingestion_pipeline.indexing.indexer import IndexingError
 from ingestion_pipeline.orchestration.pipeline import Pipeline, PipelineError
